+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 * Unreleased
+    * Update `examples/MemoryBenchmark` numbers from v0.8 to v1.2 with
+      new auto-generator scripts.
 * 1.2 (2020-10-31, TZ DB version 2020d)
     * **Potentially Breaking**: AceTime library now depends on the AceCommon
       library (https://github.com/bxparks/AceCommon) to avoid having to maintain